refactor: move input note record files into module
