+++ crates/rust-client/src/store/mod.rs
@@ -46,8 +46,8 @@ pub mod web_store;
 mod note_record;
 pub use note_record::{
     CommittedNoteState, ConsumedAuthenticatedLocalNoteState, ExpectedNoteState, InputNoteRecord,
-    NoteRecordDetails, NoteRecordError, NoteState, NoteStatus, OutputNoteRecord,
-    ProcessingAuthenticatedNoteState, ProcessingUnauthenticatedNoteState, STATE_COMMITTED,
+    InvalidNoteState, NoteRecordDetails, NoteRecordError, NoteState, NoteStatus, OutputNoteRecord,
+    ProcessingAuthenticatedNoteState, ProcessingUnauthenticatedNoteState,
     STATE_CONSUMED_AUTHENTICATED_LOCAL, STATE_CONSUMED_EXTERNAL,
     STATE_CONSUMED_UNAUTHENTICATED_LOCAL, STATE_EXPECTED, STATE_PROCESSING_AUTHENTICATED,
     STATE_PROCESSING_UNAUTHENTICATED, STATE_UNVERIFIED,
