+++ crates/rust-client/src/store/note_record/input_note_record/mod.rs
@@ -8,9 +8,15 @@ use miden_objects::{
     BlockHeader, Digest,
 };
 
-use super::{
-    states::{ExpectedNoteState, UnverifiedNoteState},
-    NoteRecordError, NoteState,
+use super::NoteRecordError;
+
+mod states;
+pub use states::{
+    CommittedNoteState, ConsumedAuthenticatedLocalNoteState, ExpectedNoteState, InvalidNoteState,
+    NoteState, ProcessingAuthenticatedNoteState, ProcessingUnauthenticatedNoteState,
+    UnverifiedNoteState, STATE_COMMITTED, STATE_CONSUMED_AUTHENTICATED_LOCAL,
+    STATE_CONSUMED_EXTERNAL, STATE_CONSUMED_UNAUTHENTICATED_LOCAL, STATE_EXPECTED,
+    STATE_PROCESSING_AUTHENTICATED, STATE_PROCESSING_UNAUTHENTICATED, STATE_UNVERIFIED,
 };
 
 // INPUT NOTE RECORD
