+++ crates/rust-client/src/store/note_record/input_note_record/states/committed.rs
