+++ crates/rust-client/src/store/note_record/mod.rs
@@ -44,17 +44,15 @@ use miden_objects::{
 
 mod input_note_record;
 mod output_note_record;
-mod states;
 
-pub use input_note_record::InputNoteRecord;
-pub use output_note_record::OutputNoteRecord;
-pub use states::{
-    CommittedNoteState, ConsumedAuthenticatedLocalNoteState, ExpectedNoteState, NoteState,
-    ProcessingAuthenticatedNoteState, ProcessingUnauthenticatedNoteState, STATE_COMMITTED,
-    STATE_CONSUMED_AUTHENTICATED_LOCAL, STATE_CONSUMED_EXTERNAL,
-    STATE_CONSUMED_UNAUTHENTICATED_LOCAL, STATE_EXPECTED, STATE_PROCESSING_AUTHENTICATED,
-    STATE_PROCESSING_UNAUTHENTICATED, STATE_UNVERIFIED,
+pub use input_note_record::{
+    CommittedNoteState, ConsumedAuthenticatedLocalNoteState, ExpectedNoteState, InputNoteRecord,
+    InvalidNoteState, NoteState, ProcessingAuthenticatedNoteState,
+    ProcessingUnauthenticatedNoteState, STATE_COMMITTED, STATE_CONSUMED_AUTHENTICATED_LOCAL,
+    STATE_CONSUMED_EXTERNAL, STATE_CONSUMED_UNAUTHENTICATED_LOCAL, STATE_EXPECTED,
+    STATE_PROCESSING_AUTHENTICATED, STATE_PROCESSING_UNAUTHENTICATED, STATE_UNVERIFIED,
 };
+pub use output_note_record::OutputNoteRecord;
 
 // NOTE STATUS
 // ================================================================================================
